[musl] Fix userspace int types and format specifiers
This also fixes two bugs in the kernel libc equivalents.

In addition, remove a number of unneeded inclusions of <sys/types.h>
throughout system, and fix some format specifiers. Also add a test of
these things (which found the bugs mentioned above).

+++ system/uapp/lsblk/main.c
@@ -139,7 +139,7 @@ static int cmd_read_blk(const char* dev, off_t offset, size_t count) {
     if (offset) {
         rc = lseek(fd, offset, SEEK_SET);
         if (rc < 0) {
-            printf("Error %zd seeking to offset %" PRId64 "\n", rc, offset);
+            printf("Error %zd seeking to offset %jd\n", rc, (intmax_t)offset);
             goto out2;
         }
     }
